+++ changelog.md
@@ -6,3 +6,8 @@
 ## [v3.02] - Layers Class  |  21/03/26
 - Added `Layers` Class in `layers.py`.
 - Created a `config.json` and `data.json` file in `Files\`.
+
+
+## [v3.03] - Trainer Class  |  25/03/26
+- Started `Trainer` Class in `trainer.py`.
+- Created `randominit()` in `utils.py`.